Sims

Sims Solver

RationalExpectations.sims — Method.

sims(G0,G1,Pi,Psi)

This function solves a rational expectations model using Sims(2000) (aka gensys). The model should be written in the following way:

\[\Gamma_0 \mathbf{x_{t+1}} = \Gamma_1 \mathbf{x_{t}} + \Psi \mathbf{\varepsilon_t} + \Pi \mathbf{ta_t}\]

The arguments of the function are:

G0 is $\Gamma_0$

G1 is $\Gamma_1$

Pi is $\Pi$

Psi is $\Psi$

Where $\Gamma_0$ can be singular. $\eta_t$ is the expectation error and $E_{t}(\nu_{t+1})=0$ by construction.

Sims`s solver does not require that we say what variables are jump variables.

Value

The function returns theta1 and theta2, such that:

\[y_t = \Theta_1 y_{t-1} + \Theta_2 \varepsilon_t\]

Note

Although the original algorithm allows cases with multiple equilibrium (sunspots), this has not been implemented thus far.

source

RationalExpectations.irf — Method.

irf(model::SimsSol,t::Int,shock)
irf(Theta1,Theta2,t,shock)

Generates the IRF from the matrices calculate using the gensys.

This function allows you to pass the matrices directly or to provide the whole model.

model receives the full output of the function sims

Theta1 is the theta1 matrix from sims

Theta2 is the theta1 matrix from sims

t is the number of periodos to be simulated

shock is the size of the shock

See also sims(G0,G1,Pi,Psi)

source

Issues

Although most implementations force to use the complex Schur, we don`t do it here (and discourage it). Here what happens when you use the schur decomposition with complex $\Gamma_0$ and $\Gamma_1$:

using Plots
using RationalExpectations

sigma = 1
phi_pi = 1.5
phi_y = 0.5/4
beta = 0.99
phi = 1
alph = 1/3
ep = 6
theta = 2/3
rho_v = 0.5

Theta = (1-alph)/(1-alph+alph*ep)
lambda = (1-theta)*((1-beta*theta)/theta)*Theta
kappa = lambda*(sigma+(phi+alph)/(1-alph))

G0 = [[1 0 0 0];[-1 1 0 0];[0 -1 sigma 1];[0 0 0 beta]]
G1 = [[rho_v 0 0 0];[0 0 phi_y phi_pi];[0 0 sigma 0];[0 0 -kappa 1]]
Psi = [1 0 0 0]'
Pi = [[0 phi_pi 0 1];[0 phi_y sigma -kappa]]'

sol_sims = sims(complex(G0),complex(G1),Pi,Psi)

resul = irf(real(sol_sims.Theta1),real(sol_sims.Theta2),12,0.25)

[ Info: Unique Solution Available
13×4 Array{Float64,2}:
 -0.0339898    -0.154404     0.019368     0.00488993
 -0.0169949    -0.00723901   0.00968398   0.00244496
 -0.00849745   -0.00361951   0.00484199   0.00122248
 -0.00424873   -0.00180975   0.00242099   0.000611241
 -0.00212436   -0.000904877  0.0012105    0.000305621
 -0.00106218   -0.000452438  0.000605249  0.00015281
 -0.000531091  -0.000226219  0.000302624  7.64051e-5
 -0.000265545  -0.00011311   0.000151312  3.82026e-5
 -0.000132773  -5.65548e-5   7.56561e-5   1.91013e-5
 -6.63863e-5   -2.82774e-5   3.7828e-5    9.55064e-6
 -3.31932e-5   -1.41387e-5   1.8914e-5    4.77532e-6
 -1.65966e-5   -7.06935e-6   9.45701e-6   2.38766e-6
 -8.29829e-6   -3.53467e-6   4.7285e-6    1.19383e-6

Let's plot the IRF of the shock to the monetary policy, that is an AR(1) with $\rho_v = 0.5$:

plot(resul[:,1], leg = false)

Solutions are welcomed. On the other hand, klein does not suffer from this issue.